Implemented a DWR ajax solution similar to how "Deploy Study" does its asynchronous persisted job.
GForge Tracker #21154
Peer reviewed by JP

+++ software/caintegrator2-war/src/gov/nih/nci/caintegrator2/web/action/study/management/EditGenomicSourceAction.java
@@ -87,10 +87,16 @@ package gov.nih.nci.caintegrator2.web.action.study.management;
 
 import gov.nih.nci.caintegrator2.application.arraydata.ArrayDataService;
 import gov.nih.nci.caintegrator2.application.arraydata.PlatformVendorEnum;
+import gov.nih.nci.caintegrator2.application.study.GenomicDataSourceConfiguration;
+import gov.nih.nci.caintegrator2.application.study.Status;
 import gov.nih.nci.caintegrator2.domain.genomic.Platform;
+import gov.nih.nci.caintegrator2.external.ServerConnectionProfile;
+import gov.nih.nci.caintegrator2.web.ajax.IGenomicDataSourceAjaxUpdater;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import org.apache.commons.lang.StringUtils;
 
 /**
  * Action called to create or edit a <code>GenomicDataSourceConfiguration</code>.
@@ -99,6 +105,7 @@ public class EditGenomicSourceAction extends AbstractGenomicSourceAction {
 
     private static final long serialVersionUID = 1L;
     
+    private IGenomicDataSourceAjaxUpdater updater;
     private ArrayDataService arrayDataService;
 
     /**
@@ -110,13 +117,75 @@ public class EditGenomicSourceAction extends AbstractGenomicSourceAction {
     }
     
     /**
+     * Cancels the creation of a genomic source to return back to study screen.
+     * @return struts string.
+     */
+    public String cancel() {
+        return SUCCESS;
+    }
+    
+    /**
      * Delete a genomic source file.
      * @return struts string.
      */
     public String delete() {
+        if (getGenomicSource() == null 
+           || !getStudyConfiguration().getGenomicDataSources().contains(getGenomicSource())) {
+            return SUCCESS;
+        }
         getStudyManagementService().delete(getStudyConfiguration(), getGenomicSource());
         return SUCCESS;
     }
+    
+    /**
+     * {@inheritDoc}
+     */
+    public String save() {
+        if (!validateSave()) {
+            return INPUT;
+        }
+        getStudyConfiguration().setStatus(Status.NOT_DEPLOYED);
+        if (getGenomicSource().getId() == null) {
+            runAsynchronousGenomicDataRetrieval(getGenomicSource());
+        } else { // Need to create a new source, delete the old one, and load the new one
+            GenomicDataSourceConfiguration newGenomicSource = createNewGenomicSource();
+            delete();
+            runAsynchronousGenomicDataRetrieval(newGenomicSource);
+        }
+        return SUCCESS;
+    }
+
+    private GenomicDataSourceConfiguration createNewGenomicSource() {
+        GenomicDataSourceConfiguration configuration = new GenomicDataSourceConfiguration();
+        ServerConnectionProfile newProfile = configuration.getServerProfile();
+        ServerConnectionProfile oldProfile = getGenomicSource().getServerProfile();
+        newProfile.setHostname(oldProfile.getHostname());
+        newProfile.setPort(oldProfile.getPort());
+        newProfile.setUsername(oldProfile.getUsername());
+        newProfile.setPassword(oldProfile.getPassword());
+        configuration.setExperimentIdentifier(getGenomicSource().getExperimentIdentifier());
+        configuration.setPlatformVendor(getGenomicSource().getPlatformVendor());
+        configuration.setPlatformName(getGenomicSource().getPlatformName());
+        return configuration;
+    }
+
+    private void runAsynchronousGenomicDataRetrieval(GenomicDataSourceConfiguration genomicSource) {
+        getDisplayableWorkspace().setCurrentStudyConfiguration(getStudyConfiguration());
+        genomicSource.setStatus(Status.PROCESSING);
+        getStudyManagementService().addGenomicSourceToStudy(getStudyConfiguration(), genomicSource);
+        updater.runJob(genomicSource);
+    }
+
+    private boolean validateSave() {
+        if (PlatformVendorEnum.AGILENT.getValue().equals(getGenomicSource().getPlatformVendor())
+                && StringUtils.isEmpty(getGenomicSource().getPlatformName())) {
+            addFieldError("genomicSource.platformName", "Platform name is required for Agilent");
+            return false;
+        }
+        return true;
+    }
+    
+    
     
     /**
      * @return all platform names
@@ -144,6 +213,20 @@ public class EditGenomicSourceAction extends AbstractGenomicSourceAction {
      */
     public void setArrayDataService(ArrayDataService arrayDataService) {
         this.arrayDataService = arrayDataService;
+    }
+
+    /**
+     * @return the updater
+     */
+    public IGenomicDataSourceAjaxUpdater getUpdater() {
+        return updater;
+    }
+
+    /**
+     * @param updater the updater to set
+     */
+    public void setUpdater(IGenomicDataSourceAjaxUpdater updater) {
+        this.updater = updater;
     }
        
 }
